+++ app/src/main/java/uk/org/openseizuredetector/MainActivity2.java
@@ -318,8 +318,10 @@ public class MainActivity2 extends AppCompatActivity {
                     return new FragmentHrAlg();
                 case 2:
                     return new FragmentSystem();
-                //case 3:
-                //    return new FragmentBatt();
+                case 3:
+                    return new FragmentWatchSig();
+                case 4:
+                    return new FragmentBatt();
                 //case 4:
                 //    return new FragmentDataSharing();
 
@@ -331,7 +333,7 @@ public class MainActivity2 extends AppCompatActivity {
 
         @Override
         public int getItemCount() {
-            return 3;
+            return 5;
         }
     }