+++ app/src/main/java/uk/org/openseizuredetector/SdServer.java
@@ -821,15 +821,19 @@ public class SdServer extends Service implements SdDataReceiver {
         // flag.
         if (mFaultTimerCompleted) {
             faultWarningBeep();
-            // Re-start the data source to see if that fixes it
-            Log.w(TAG,"FAULT - stopping data source");
-            mSdDataSource.stop();
-            mHandler.postDelayed(new Runnable() {
-                public void run() {
-                    Log.w(TAG,"FAULT - restarting data source");
-                    mSdDataSource.start();
-                }
-            }, 10000);
+            // Disable the data-source re-start for now because it was messing up BLE2 data source by ending up with multiple
+            // notifications for the same data when it reconnects.
+            if (false) {
+                // Re-start the data source to see if that fixes it
+                Log.w(TAG, "FAULT - stopping data source");
+                mSdDataSource.stop();
+                mHandler.postDelayed(new Runnable() {
+                    public void run() {
+                        Log.w(TAG, "FAULT - restarting data source");
+                        mSdDataSource.start();
+                    }
+                }, 10000);
+            }
         } else {
             startFaultTimer();
             Log.v(TAG, "onSdDataFault() - starting Fault Timer");